+++ R/nchars.R
@@ -2,8 +2,8 @@
 #' 
 #' @param x Object to process.
 #' @param progress A `logical` value.
-#' @param regexCharsToKeep if not NULL, a regex
-#' @param toLower whether to lower tokens
+#' @param char_regex If not NULL, a regular expression - defaults to "[a-zA-Z]".
+#' @param lowercase whether to lower tokens
 #' @param mc logical
 #' @param ... Argument passed into `blapply()`.
 #' @param decreasing logical, passed into order call 
@@ -13,8 +13,8 @@ setGeneric("nchars", function(x, ...) standardGeneric("nchars"))
 
 
 #' @param p_attribute the p-attribute
-#' @param regexCharsToKeep if NULL, counts for all charactrs will be returned, else a regex indicating which characters to include in the counting
-#' @param toLower whether to lower tokens
+#' @param char_regex if NULL, counts for all charactrs will be returned, else a regex indicating which characters to include in the counting
+#' @param lowercase whether to lower tokens
 #' @param mc logical
 #' @param decreasing logical, passed into order call 
 #' @exportMethod nchars
@@ -37,25 +37,25 @@ setGeneric("nchars", function(x, ...) standardGeneric("nchars"))
 #' corpus("REUTERS") %>%
 #'   split(s_attribute = "id") %>%
 #'   nchars()
-setMethod("nchars", "partition", function(x, p_attribute = "word", regexCharsToKeep = "[a-zA-Z]", toLower = TRUE, decreasing = TRUE){
+setMethod("nchars", "partition", function(x, p_attribute = "word", char_regex = "[a-zA-Z]", lowercase = TRUE, decreasing = TRUE){
   .Object <- x
   charSoup <- get_token_stream(.Object, p_attribute = p_attribute, collapse = "")
-  if (isTRUE(toLower)) charSoup <- tolower(charSoup)
+  if (isTRUE(lowercase)) charSoup <- tolower(charSoup)
   charCount <- table(unlist(strsplit(charSoup, "")))
-  if(!is.null(regexCharsToKeep)){
-    charCount <- charCount[grep(regexCharsToKeep, names(charCount))]
+  if(!is.null(char_regex)){
+    charCount <- charCount[grep(char_regex, names(charCount))]
   }
   y <- charCount[order(charCount, decreasing = decreasing)]
   setNames(as.integer(y), names(y))
 })
 
 #' @rdname nchars
-setMethod("nchars", "subcorpus", function(x, p_attribute = "word", regexCharsToKeep = "[a-zA-Z]", toLower = TRUE, decreasing = TRUE){
+setMethod("nchars", "subcorpus", function(x, p_attribute = "word", char_regex = "[a-zA-Z]", lowercase = TRUE, decreasing = TRUE){
   nchars(
     x = as(x, "partition"),
     p_attribute = p_attribute,
-    regexCharsToKeep = regexCharsToKeep,
-    toLower = toLower,
+    char_regex = char_regex,
+    lowercase = lowercase,
     decreasing = decreasing
   )
 })
@@ -92,6 +92,8 @@ setMethod("nchars", "subcorpus_bundle", function(x, decreasing = TRUE, mc = FALS
   )
 })
 
+
+#' @param verbose Whether to output progress messages.
 #' @param sample An `integer` or `numeric` value defining the number of sample
 #'   tokens extracted from the (entirely decoded) token stream to be evaluated.
 #' @rdname nchars
@@ -101,24 +103,39 @@ setMethod("nchars", "subcorpus_bundle", function(x, decreasing = TRUE, mc = FALS
 #' library(polmineR)
 #' use("RcppCWB")
 #' n <- corpus("REUTERS") %>% nchars(sample = 4000)
-setMethod("nchars", "corpus", function(x, p_attribute = "word", toLower = TRUE, sample = 5000000L, regexCharsToKeep = "[a-zA-Z]", decreasing = TRUE, mc = FALSE, progress = TRUE){
+setMethod("nchars", "corpus", function(x, p_attribute = "word", lowercase = TRUE, sample = 5000000L, char_regex = "[a-zA-Z]", decreasing = TRUE, verbose = TRUE){
   # optime get_token_stream(), use it here and callNextMethod() for 
   # partition- and subcorpus-method
+  cli::cli_progress_step(msg = "decode token stream")
   tokens <- decode(
     0L:(x@size - 1L),
     corpus = x,
     boost = TRUE,
     p_attributes = p_attribute
   )
-  if (isTRUE(toLower)) tokens <- tolower(tokens)
+  
+  if (isTRUE(lowercase)){
+    cli::cli_progress_step(msg = "lowercase token stream")
+    tokens <- tolower(tokens)
+  }
+  
   if (is.numeric(sample)){
+    cli::cli_progress_step(msg = "draw sample")
     tokens <- sample(tokens, size = sample)
     gc()
   }
   
-  n <- table(unlist(strsplit(tokens, "")))
-  if(!is.null(regexCharsToKeep)) n <- n[grep(regexCharsToKeep, names(n))]
+  cli::cli_progress_step(msg = "split into characters")
+  chars <- unlist(strsplit(tokens, "", fixed = TRUE), recursive = FALSE)
   
+  cli::cli_progress_step(msg = "tabulate")
+  n <- table(chars)
+  
+  cli::cli_progress_step(msg = "apply regular expression")
+  if(!is.null(char_regex)) n <- n[grep(char_regex, names(n))]
+  
+  cli::cli_progress_step(msg = "order result")
   y <- n[order(n, decreasing = decreasing)]
+  
   setNames(as.integer(y), names(y))
 })